+++ includes/class-nuvei-logger.php
@@ -19,11 +19,9 @@ class Nuvei_Logger
      */
 	public static function write( $data, $message = '', $log_level = 'INFO', $span_id = '')
     {
-		$plugin_data        = get_plugin_data(plugin_dir_path(NUVEI_PLUGIN_FILE) . 'index.php');
         $nuvei_gw           = WC()->payment_gateways->payment_gateways()[NUVEI_GATEWAY_NAME];
-		$save_logs          = $nuvei_gw->get_option('save_logs');
+        $save_logs          = $nuvei_gw->get_option('save_logs');
 		$save_single_log    = $nuvei_gw->get_option('save_single_log');
-		$test_mode          = $nuvei_gw->get_option('test');
         
 		if (!is_dir(NUVEI_LOGS_DIR)) {
 			return;
@@ -31,6 +29,10 @@ class Nuvei_Logger
 		if ('no' == $save_logs && 'no' == $save_single_log) {
 			return;
 		}
+        
+        $plugin_data    = get_plugin_data(plugin_dir_path(NUVEI_PLUGIN_FILE) . 'index.php');
+		$test_mode      = $nuvei_gw->get_option('test');
+		$maskUserData   = $nuvei_gw->get_option('mask_user_data');
 		
         $beauty_log = ('yes' == $test_mode) ? true : false;
         $tab        = '    '; // 4 spaces
@@ -75,6 +77,10 @@ class Nuvei_Logger
         }
         
         if(is_array($data)) {
+            if (1 == $maskUserData) {
+                $data = self::mask_data($data);
+            }
+
             // paymentMethods can be very big array
             if(!empty($data['paymentMethods'])) {
                 $exception = json_encode($data);
@@ -139,4 +145,79 @@ class Nuvei_Logger
             );
         }
 	}
+    
+    /**
+     * Mask data for the log.
+     * 
+     * @param array $data
+     */
+    private static function mask_data($data)
+    {
+        // mask the IP address
+        if (!empty($data[LOG_REQUEST_PARAMS]['deviceDetails'])) {
+            $data[LOG_REQUEST_PARAMS]['deviceDetails']['ipAddress']
+                = rtrim(long2ip(ip2long($data[LOG_REQUEST_PARAMS]['deviceDetails']['ipAddress']) & (~255)),"0")."x";
+        }
+        
+        // mask the shipping details
+        if (!empty($data[LOG_REQUEST_PARAMS]['shippingAddress'])) {
+            $data[LOG_REQUEST_PARAMS]['shippingAddress']
+                = self::mask_string($data[LOG_REQUEST_PARAMS]['shippingAddress']);
+        }
+        
+        // mask the billing details
+        if (!empty($data[LOG_REQUEST_PARAMS]['billingAddress'])) {
+            $data[LOG_REQUEST_PARAMS]['billingAddress']
+                = self::mask_string($data[LOG_REQUEST_PARAMS]['billingAddress']);
+        }
+        
+        if (!empty($data['billingAddress'])) {
+            $data['billingAddress'] = self::mask_string($data['billingAddress']);
+        }
+        
+        // mask the user details
+        if (!empty($data[LOG_REQUEST_PARAMS]['userDetails'])) {
+            $data[LOG_REQUEST_PARAMS]['userDetails']
+                = self::mask_string($data[LOG_REQUEST_PARAMS]['userDetails']);
+        }
+        
+        // other, like responses
+        $data = self::mask_string($data, false);
+
+        return $data;
+    }
+    
+    /**
+     * A help function to mask user data.
+     * 
+     * @param type $array           Original array.
+     * @param bool $is_user_data    If this is no user data, do not mask all fields.
+     * 
+     * @return array $array         Array with masked strings.
+     */
+    private static function mask_string($array, $is_user_data = true)
+    {
+        foreach ($array as $key => $value) {
+            // mask the email
+            if (in_array($key, ['email', 'userTokenId']) && !empty($value)) {
+                $part = substr($value, 4);
+                $array[$key] = '****' . $part;
+                continue;
+            }
+            
+            // mask the names
+            if (strpos(strtolower($key), 'name')) {
+                $array[$key] = substr($value, 0, 1) . '****';
+                continue;
+            }
+            
+            // mask the others
+            if ($is_user_data && !is_numeric($key) && !in_array($key, ['city', 'country'])) {
+                $array[$key] = '****';
+            } 
+        }
+        
+        return $array;
+    }
+    
 }